this is the main file. there may be more coming.
the code is really easy, just give the script an navmesh agent and since the script has a boolean for that you just hook that boolean up to where it will go. this can be applyed to really anything if you want it to. if you want a tutorial please download this, the more downloads the more updates and tutorials i will do.